requirements/requirements-main.txt
# Vocal10n — Main Dependencies (Python 3.11, venv_main)
# Covers: STT + LLM + UI + Pipeline + OBS overlay

# === UI ===
PySide6>=6.6.0

# === Core ===
pyyaml>=6.0
numpy>=1.24.0
nvidia-ml-py>=12.0.0
requests>=2.31.0
psutil>=5.9.0
aiohttp>=3.9.0

# === STT (FasterWhisper) ===
faster-whisper>=1.0.0
sounddevice>=0.4.6
soundfile>=0.12.1
scipy>=1.11.0

# === Speaker Diarization (optional — for speaker tagging) ===
pyannote-audio>=3.3.0

# === Chinese text processing ===
opencc-python-reimplemented>=0.1.7
pypinyin>=0.49.0

# === LLM (Qwen3-4B via llama-cpp-python) ===
# Install with CUDA support:
#   pip install llama-cpp-python --extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cu121
llama-cpp-python>=0.2.77

# === TTS (Qwen3-TTS — optional, alternative to GPT-SoVITS) ===
# Requires Python 3.12 ideally; works in 3.11 with some constraints.
# Install flash-attn separately for lower VRAM usage:
#   pip install flash-attn --no-build-isolation
qwen-tts>=0.1.0

# === RAG (optional — for large glossaries) ===
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# === OBS overlay server ===
flask>=3.0.0
flask-cors>=4.0.0
